less fire damage

+++ backend/src/game/systems/monsters_wave.rs
@@ -107,7 +107,7 @@ fn generate_all_monsters_specs(
 fn generate_monster_specs(bp_specs: &BaseMonsterSpecs, world_state: &WorldState) -> MonsterSpecs {
     let mut monster_specs = MonsterSpecs::init(bp_specs);
     let exp_factor = increase_factors::exponential(
-        world_state.area_level as f64,
+        world_state.area_level,
         increase_factors::MONSTER_INCREASE_FACTOR,
     );
     let lin_factor = increase_factors::linear(world_state.area_level as f64);
